Move camera and other screens
