+++ src/screen/index.js
@@ -1,0 +1,3 @@
+export { default as AccountScreen } from "./account";
+export { default as CameraScreen } from "./camera";
+export { default as SettingsScreen } from "./settings";
